Added check if IP address reported by DNS match address in cache file
It happen to me a couple of times, that my real IP address was in cache file, but DuiaDNS was reporting different IP.
In this patch I retrive IP address for hostname from DuiaDNS using dig or form generec DNS server using wget, if dig is not available.
If retrived address doesn't match my real IP and cache file is older than amount of time specified in duiadns_propagation_timeout, update is forced.